\name{GoogleDataSearch-package}
\alias{GoogleDataSearch-package}
\alias{GoogleDataSearch}
\docType{package}
\title{
Dataframe building from Google Maps queries to allow for Spatial Economic Analysis
}
\description{
Provide functions that create data sets from different types of queries that are parsed and downloaded from Google Maps.
}
\details{
\tabular{ll}{
Package: \tab GoogleDataSearch\cr
Imports: \tab XML, RCurl \cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-10-15\cr
License: \tab GNU Public License\cr
}
This package provides three functions that create a \code{data.frame} object from Google Maps queries: 1) Google Maps Query of a place in general 2) Google Maps Query within a radious and categories,and 3) a Matrix of Distances from origin and destitation coordinates in lat and long format.
}
\author{
Esteban Lopez
\maintainer{Esteban Lopez <lopezoch@illinois.edu>
}}

\references{
For more information go to: \href{https://developers.google.com/places/web-service}{https://developers.google.com/places/web-service}
}

\keyword{Google Maps, data.frame}
\seealso{

}
\examples{

}
