\name{GoogleDataSearch-package}
\alias{GoogleDataSearch-package}
\alias{GoogleDataSearch}
\docType{package}
\title{
Dataframe building from Google Maps queries to allow for Spatial Economic Analysis
GoogleDataSearch
}
\description{
Provide functions that create data sets from different types of queries as it they were made on Google Maps.
}
\details{
\tabular{ll}{
Package: \tab GoogleDataSearch\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-15\cr
License: \tab GNU Public License\cr
}
This package provides two functions: 1) Query Google Maps in general 2) Query Google Maps within a radious and categories.
}
\author{
Esteban Lopez
Maintainer: Esteban Lopez <lopezoch@illinois.edu>
}
\references{

}

\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-plotGoogleMaps]{<pkg>}}
}
\examples{

}
